[go] Optimize pagination
The goal of this commit is to reduce the number of pages read by the
pager.
Before, the pager would read until the page length was equal to 0.
But, if the page length is less than the limit it already means that
there is no more results after the current page.
This commit can significantly reduce the number of read operations.